+++ tests/TestCase/View/Helper/AuthHelperTest.php
@@ -20,6 +20,17 @@ use MeTools\TestSuite\HelperTestCase;
 class AuthHelperTest extends HelperTestCase
 {
     /**
+     * Internal method to write auth data on session
+     * @param array $data Data you want to write
+     * @return void
+     */
+    protected function writeAuthOnSession(array $data = [])
+    {
+        $this->Helper->getView()->getRequest()->getSession()->write('Auth.User', $data);
+        $this->Helper->initialize([]);
+    }
+
+    /**
      * Tests for `hasId()` method
      * @test
      */
@@ -27,7 +38,7 @@ class AuthHelperTest extends HelperTestCase
     {
         $this->assertFalse($this->Helper->hasId(1));
 
-        $this->Helper->initialize(['user' => ['id' => 1]]);
+        $this->writeAuthOnSession(['id' => 1]);
         $this->assertTrue($this->Helper->hasId(1));
         $this->assertTrue($this->Helper->hasId([1, 2]));
         $this->assertFalse($this->Helper->hasId(2));
@@ -42,10 +53,10 @@ class AuthHelperTest extends HelperTestCase
     {
         $this->assertFalse($this->Helper->isFounder());
 
-        $this->Helper->initialize(['user' => ['id' => 1]]);
+        $this->writeAuthOnSession(['id' => 1]);
         $this->assertTrue($this->Helper->isFounder());
 
-        $this->Helper->initialize(['user' => ['id' => 2]]);
+        $this->writeAuthOnSession(['id' => 2]);
         $this->assertFalse($this->Helper->isFounder());
     }
 
@@ -57,7 +68,7 @@ class AuthHelperTest extends HelperTestCase
     {
         $this->assertFalse($this->Helper->isGroup('admin'));
 
-        $this->Helper->initialize(['user' => ['group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         $this->assertTrue($this->Helper->isGroup('admin'));
         $this->assertTrue($this->Helper->isGroup(['admin', 'manager']));
         $this->assertFalse($this->Helper->isGroup('manager'));
@@ -72,7 +83,7 @@ class AuthHelperTest extends HelperTestCase
     {
         $this->assertFalse($this->Helper->isLogged());
 
-        $this->Helper->initialize(['user' => ['id' => 1]]);
+        $this->writeAuthOnSession(['id' => 1]);
         $this->assertTrue($this->Helper->isLogged());
     }
 
@@ -85,7 +96,7 @@ class AuthHelperTest extends HelperTestCase
         $this->assertEmpty($this->Helper->user());
         $this->assertEmpty($this->Helper->user('id'));
 
-        $this->Helper->initialize(['user' => ['id' => 1, 'group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['id' => 1, 'group' => ['name' => 'admin']]);
         $this->assertEquals(['id' => 1, 'group' => ['name' => 'admin']], $this->Helper->user());
         $this->assertEquals(1, $this->Helper->user('id'));
         $this->assertEmpty($this->Helper->user('noExistingKey'));