+++ tests/TestCase/View/Helper/MenuHelperTest.php
@@ -21,6 +21,17 @@ use MeTools\View\Helper\HtmlHelper;
 class MenuHelperTest extends HelperTestCase
 {
     /**
+     * Internal method to write auth data on session
+     * @param array $data Data you want to write
+     * @return void
+     */
+    protected function writeAuthOnSession(array $data = [])
+    {
+        $this->Helper->getView()->getRequest()->getSession()->write('Auth.User', $data);
+        $this->Helper->Auth->initialize([]);
+    }
+
+    /**
      * Internal method to build links
      * @param array $links Links
      * @return array
@@ -55,7 +66,7 @@ class MenuHelperTest extends HelperTestCase
             '<a href="/me-cms/admin/posts-tags" title="List tags">List tags</a>',
         ];
         foreach (['manager', 'admin'] as $name) {
-            $this->Helper->Auth->initialize(['user' => ['group' => compact('name')]]);
+            $this->writeAuthOnSession(['group' => compact('name')]);
             $this->assertEquals($expected, $this->buildLinks($this->Helper->posts()[0]));
         }
     }
@@ -82,7 +93,7 @@ class MenuHelperTest extends HelperTestCase
             '<a href="/me-cms/admin/pages/index-statics" title="List static pages">List static pages</a>',
         ];
         foreach (['manager', 'admin'] as $name) {
-            $this->Helper->Auth->initialize(['user' => ['group' => compact('name')]]);
+            $this->writeAuthOnSession(['group' => compact('name')]);
             $this->assertEquals($expected, $this->buildLinks($this->Helper->pages()[0]));
         }
     }
@@ -106,7 +117,7 @@ class MenuHelperTest extends HelperTestCase
         $this->assertEquals(['icon' => 'camera-retro'], $options);
 
         foreach (['manager', 'admin'] as $name) {
-            $this->Helper->Auth->initialize(['user' => ['group' => compact('name')]]);
+            $this->writeAuthOnSession(['group' => compact('name')]);
             $this->assertEquals($expected, $this->buildLinks($this->Helper->photos()[0]));
         }
     }
@@ -124,7 +135,7 @@ class MenuHelperTest extends HelperTestCase
             '<a href="/me-cms/admin/banners/upload" title="Upload banners">Upload banners</a>',
         ];
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'manager']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
         list($links, $title, $options) = $this->Helper->banners();
         $this->assertEquals($expected, $this->buildLinks($links));
         $this->assertEquals('Banners', $title);
@@ -132,7 +143,7 @@ class MenuHelperTest extends HelperTestCase
 
         $expected[] = '<a href="/me-cms/admin/banners-positions" title="List positions">List positions</a>';
         $expected[] = '<a href="/me-cms/admin/banners-positions/add" title="Add position">Add position</a>';
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         list($links) = $this->Helper->banners();
         $this->assertEquals($expected, $this->buildLinks($links));
     }
@@ -150,7 +161,7 @@ class MenuHelperTest extends HelperTestCase
             '<a href="/me-cms/admin/users/add" title="Add user">Add user</a>',
         ];
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'manager']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
         list($links, $title, $options) = $this->Helper->users();
         $this->assertEquals($expected, $this->buildLinks($links));
         $this->assertEquals('Users', $title);
@@ -158,7 +169,7 @@ class MenuHelperTest extends HelperTestCase
 
         $expected[] = '<a href="/me-cms/admin/users-groups" title="List groups">List groups</a>';
         $expected[] = '<a href="/me-cms/admin/users-groups/add" title="Add group">Add group</a>';
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         list($links) = $this->Helper->users();
         $this->assertEquals($expected, $this->buildLinks($links));
     }
@@ -171,10 +182,10 @@ class MenuHelperTest extends HelperTestCase
     {
         $this->assertNull($this->Helper->backups());
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'manager']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
         $this->assertNull($this->Helper->backups());
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         list($links, $title, $options) = $this->Helper->backups();
         $this->assertEquals([
             '<a href="/me-cms/admin/backups" title="List backups">List backups</a>',
@@ -192,7 +203,7 @@ class MenuHelperTest extends HelperTestCase
     {
         $this->assertNull($this->Helper->systems());
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'manager']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
         list($links, $title, $options) = $this->Helper->systems();
         $this->assertEquals([
             '<a href="/me-cms/admin/systems/tmp-viewer" title="Temporary files">Temporary files</a>',
@@ -203,7 +214,7 @@ class MenuHelperTest extends HelperTestCase
         $this->assertEquals('System', $title);
         $this->assertEquals(['icon' => 'wrench'], $options);
 
-        $this->Helper->Auth->initialize(['user' => ['group' => ['name' => 'admin']]]);
+        $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         $this->assertEquals([
             '<a href="/me-cms/admin/systems/tmp-viewer" title="Temporary files">Temporary files</a>',
             '<a href="/me-cms/admin/systems/checkup" title="System checkup">System checkup</a>',